pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "gitopsgui"
version = "0.1.0"
description = "GitOps management API and GUI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "gitpython>=3.1.43",
    "PyGithub>=2.3.0",
    "kubernetes>=30.1.0",
    "pydantic>=2.7.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
